requirements.txt
crewai>=0.28.0
pytube>=15.0.0
openai>=1.5.0
langchain>=0.0.267
langchain-openai>=0.0.2
python-dotenv>=1.0.0
whisper>=1.1.10
pydub>=0.25.1
ffmpeg-python>=0.2.0 